+++ packages/drizzle/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+## 0.8.5 (2025-04-28)
+
 - Feat: Enhanced the drizzleSilk function to accept configuration options, improving flexibility in schema generation.
 - Feat: add `onConflictDoUpdate` and `onConflictDoNothing` to insert mutation for Postgres and SQLite